Make create_library_config fall back to system_libs when libs is not populated
When building boost against system packages (see https://github.com/conan-io/conan/issues/14566) the `libs` list is usually empty. Instead, the libraries are put into `system_libs`. While this procedure is not fully defined yet, it seems to be common among many of the other recipes in Conan Center. Mosquitto for example fails to build against a zlib system package when it puts `z` in both `libs` and `system_libs`. It also seems logical that a library cannot be in `libs` and `system_libs` at the same time. The proposed fallback fixes the build failure against a zlib system package.